Fixes a bug in header compaction for empty containers; adds testing for all writer option combinations over all good ion-tests files.
